#140: QuizProjection uses GuavaCache to store quizzes
