+++ quiz-server/src/test/kotlin/team/undefined/quiz/core/QuizProjectionConcurrencyTest.kt
@@ -21,7 +21,7 @@ internal class QuizProjectionConcurrencyTest {
             val quiz = Quiz(name = "Awesome Quiz")
 
             val eventBus = EventBus()
-            val quizProjection = QuizProjection(eventBus, Mockito.mock(QuizStatisticsProvider::class.java), Mockito.mock(EventRepository::class.java), UndoneEventsCache())
+            val quizProjection = DefaultQuizProjection(eventBus, Mockito.mock(QuizStatisticsProvider::class.java), Mockito.mock(EventRepository::class.java), UndoneEventsCache())
 
             val observedQuiz = AtomicReference<Quiz>()
             quizProjection.observeQuiz(quiz.id)
