+++ quiz-server/src/test/kotlin/team/undefined/quiz/core/QuizProjectionForceEmitCommandTest.kt
@@ -20,7 +20,7 @@ internal class QuizProjectionForceEmitCommandTest {
                 .thenReturn(Flux.just(QuizCreatedEvent(quiz.id, quiz, quiz.timestamp)))
 
         val eventBus = EventBus()
-        val quizProjection = QuizProjection(eventBus, mock(QuizStatisticsProvider::class.java), eventRepository, UndoneEventsCache())
+        val quizProjection = DefaultQuizProjection(eventBus, mock(QuizStatisticsProvider::class.java), eventRepository, UndoneEventsCache())
 
         val observedQuiz = AtomicReference<Quiz>()
         quizProjection.observeQuiz(quiz.id)
