+++ quiz-server/src/test/kotlin/team/undefined/quiz/core/QuizProjectionQuizReloadCommandTest.kt
@@ -41,7 +41,7 @@ internal class QuizProjectionQuizReloadCommandTest {
                 ))
 
         val eventBus = EventBus()
-        val quizProjection = QuizProjection(eventBus, QuizStatisticsProvider(eventRepository), eventRepository, UndoneEventsCache())
+        val quizProjection = DefaultQuizProjection(eventBus, QuizStatisticsProvider(eventRepository), eventRepository, UndoneEventsCache())
 
         val observedQuiz = AtomicReference<Quiz>()
         quizProjection.observeQuiz(quizId)
